refactored to remove useEffect and make form uncontrolled
