+++ src/components/Details.jsx
@@ -1,6 +1,6 @@
 import { useParams } from "react-router-dom";
 import { useQuery } from "@tanstack/react-query";
-import fetchPet from "./fetchPet";
+import fetchPet from "../fetchRequests/fetchPet";
 
 const Details = () => {
   const { id } = useParams();
